fix(debt): resolve Week 22 test and lint failures
- Fix tensor_to_f16 to handle F16 dtypes (not just F32) for candle_bridge roundtrip
- Run cargo clippy --fix across all crates (273 → 31 warnings remaining)
- Export mistralrs_backend module publicly for integration tests
- Update mistralrs_integration test imports to use new public API paths

Week 22 debt work complete. Remaining: batched generation spike.

+++ pesti-safetensors/src/gguf_converter.rs
@@ -787,7 +787,7 @@ fn extract_tensor_bytes(
     tensor: &GgufTensorInfo,
 ) -> Result<Vec<u8>, GgufConvertError> {
     let mut file = std::fs::File::open(gguf_path)?;
-    let offset = tensor.offset as u64;
+    let offset = tensor.offset;
     let size = tensor.stored_size().unwrap_or(0) as usize;
 
     file.seek(std::io::SeekFrom::Start(offset))?;
